+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "Project Timer" extension will be documented in this file.
 
+# [Unreleased]
+### Fixed
+- Allow vertical scroll bar in statistics page to satisfied low resolution screens.
+
 # [0.3.2] - 2026-2-24
 ### Added
 - Error's message and stack will be logged to VS Code `Output` panel.