+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the "Project Timer" extension will be documented in this file.
 
-# [Unreleased]
+# [0.4.0] - 2026-2-24
 ### Added
 - Add new config entry `project-timer.synchronization.syncedProjects` to manage all synced projects.
 - Add new command: `project-timer.disableSyncForProject` and `project-timer.enableSyncForProject`.